rxs1052-main/backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg==3.1.13
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
